Fix dumping of readv'd data.
For readv, strace passes on all the iovec buffers (not just the part
which has actually been filled by readv), so dumping the whole stuff
would also dump a lot of junk.

Therefore we read out the message length from the header and dump
just as much as the length.